refactor(stop_filter): prefix package and namespace with autoware
